requirements.txt
qdrant-client>=1.14.2
fastembed>=0.7.0
python-dotenv>=1.0.0
google-genai
groq
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
